current fastest

+++ 2023-03/rust/src/main.rs
@@ -164,7 +164,7 @@ const LINES: usize = include_str!("../input.txt").as_bytes().len() / LINE_WIDTH;
 fn speedy_part_2(input: &str) -> usize {
     let mut input = input.as_bytes();
 
-    let mut mat = [[0; LINE_WIDTH]; LINES];
+    let mut mat = [[0u16; LINE_WIDTH]; LINES];
     let mut line = 0;
     let mut col = 0;
 
@@ -186,7 +186,7 @@ fn speedy_part_2(input: &str) -> usize {
             c @ b'0'..=b'9' => {
                 number_mode = true;
                 incr_num!(val, c);
-                mat[line][col] = number_values.len();
+                mat[line][col] = number_values.len() as u16;
                 col += 1;
             }
             b'\n' => {
@@ -236,7 +236,7 @@ fn speedy_part_2(input: &str) -> usize {
                 }
             }
 
-            Some(number_values[seen[0]] * number_values[seen[1]])
+            Some(number_values[seen[0] as usize] * number_values[seen[1] as usize])
         })
         .sum()
 }
